requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
prometheus-client==0.17.1
prometheus-flask-exporter==0.22.4
protobuf==4.21.12
Werkzeug==3.0.0
